I updated download_from_https_async function logic of downloading a file

## first improvement

previously, it downloads file by 8192 chunk size, this becomes too many chunks and it takes a bit longer time to request HTTP everytime.

New approarch is to divide file to 10 chunks, and it downloads chunks with max 5 workers concurrently.

I tested in small area of Kigali

- before changing

```
[05/07/25 08:02:22] INFO     STAC Item search: 3.04833722114563 seconds                                                                                                    stac.py:613
[05/07/25 08:05:53] INFO     STAC Item download: 210.75929975509644 seconds                                                                                                stac.py:677
[05/07/25 08:05:55] INFO     Download completed: 216.37378478050232 seconds   
```

- after changing

```
[05/07/25 09:29:11] INFO     STAC Item search: 4.000098705291748 seconds                                                                                                   stac.py:669
[05/07/25 09:32:02] INFO     STAC Item download: 170.78534173965454 seconds                                                                                                stac.py:733
[05/07/25 09:32:05] INFO     Download completed: 177.44660878181458 seconds 
```

it improves 22% faster now.

## second improvement

- after switching to gdal.warp for reprojecting

```
[05/07/25 11:18:11] INFO     STAC Item search: 3.2910964488983154 seconds                                                                                                    stac.py:739
[05/07/25 11:20:56] INFO     Download completed: 168.94079303741455 seconds  
```

- crop downloaded item by gdal.warp in tile level, and use VRT directly for prediction

```
[05/07/25 11:29:45] INFO     STAC Item search: 4.108993053436279 seconds                                                                                                   stac.py:670
[05/07/25 11:31:58] INFO     Download completed: 136.66442918777466 seconds  
```

Now approximately 58% becomes faster than current version

## Last version

I realized the current version of using https streaming is not bad. but setting chunk_size is not such effective according to the below stackoverflow. and the response time is not such different with/without chunk_size.

https://stackoverflow.com/questions/70102869/python-httpx-stream-data-asynchronously

```
[05/07/25 13:26:06] INFO     STAC Item search: 4.307923078536987 seconds                                                                                                   stac.py:606
[05/07/25 13:29:01] INFO     Download completed: 179.1740288734436 seconds   
```

This approarch is a bit slower than second version, but the code is simpler. so I prefer using this version for ease of maintenance