@JinIgarashi 2025-05-08
I tried to switch current logic to gdal.warp to directly download jp2 and reproject it. It took 477 seconds for the same area above. That is 2.2 times slower than existing version.

```
[05/08/25 08:32:13] INFO     STAC Item search: 3.186851739883423 seconds                                                                                              stac.py:644
[05/08/25 08:40:07] INFO     Download completed: 477.56029510498047 seconds                                                                                           stac.py:727
```