@JinIgarashi 2025-05-08
I add 1km buffer for project polygons and make it simplify like below. So, prediction model can compute better for the edge of project area.

<img width="623" alt="image" src="https://github.com/user-attachments/assets/bf08b6bf-a68a-4a6c-85e6-2c26208ca504" />